+++ lib/src/chart_model.dart
@@ -5,6 +5,11 @@ library highcharts.src.chart_model;
 
 import 'dart:collection' show LinkedHashMap;
 
+enum SeriesStyle { 
+  color, borderRadius, fillOpacity, size, innerSize, 
+  markerLineColor, markerSymbol, markerFillColor, markerLineWidth // marker
+}
+
 abstract class ChartModel<S extends Comparable, C extends Comparable> {
 
   /**
@@ -17,46 +22,191 @@ abstract class ChartModel<S extends Comparable, C extends Comparable> {
    */
   List<C> get categories;
 
+  // Returns vlaue of provided [series] style
+  dynamic getSeriesStyle(S series, SeriesStyle name);
+  // Sets value of provided [series] style
+  void setSeriesStyle(S series, SeriesStyle name, dynamic value);
+
   /**
    * Clear all data
    */
   void clear();
 }
 
-abstract class SingleValueCategoryModel extends ChartModel {
+abstract class AbstractChartModel<S extends Comparable, C extends Comparable> extends ChartModel<S, C> {
+  final _seriesStyle = <S, Map<SeriesStyle, dynamic>>{};
+
+  // Returns vlaue of marker of provided [series]
+  @override
+  dynamic getSeriesStyle(S series, SeriesStyle name) {
+    final kv = _seriesStyle[series];
+    return kv != null ? kv[name] : null;
+  }
+  // Sets value of marker of provided [series]
+  @override
+  void setSeriesStyle(S series, SeriesStyle name, dynamic value) {
+    final kv = _seriesStyle[series] ?? <SeriesStyle, dynamic>{};
+    kv[name] = value;
+    _seriesStyle[series] = kv;
+  }
+
+  /**
+   * Clear all data
+   */
+  @override
+  void clear() {
+    _seriesStyle.clear();
+  }
+}
+
+abstract class SingleValueCategoryModel<C extends Comparable> extends AbstractChartModel<String, C> {
 
   /**
    * Get categories as a collection.
    */
-  List<String> get categories;
+  List<C> get categories;
 
   /**
    * Get value of the specified category.
    * [category] the pie category.
    */
-  num getValue(String category);
+  num getValue(C category);
 
   /**
    * add or update the value of a specified category.
    * [category] the pie category.
    * [value] the pie value.
    */
-  void setValue(String category, num value);
+  void setValue(C category, num value);
+
+  // Sets color
+  String getValueColor(C category);
+
+  // Sets color
+  void setValueColor(C category, String color);
+
+}
+
+class DefaultSingleValueCategoryModel<C extends Comparable> extends SingleValueCategoryModel<C> {
+  final _data = <C, num>{};
+  final _color = <C, String>{};
+
+  /**
+   * Get all series as a collection.
+   */
+  @override
+  List<String> get series {
+    return <String>[];
+  }
+
+  /**
+   * Get categories as a collection.
+   */
+  @override
+  List<C> get categories => _data.keys.toList();
+
+  /**
+   * Get value of the specified category.
+   * [category] the pie category.
+   */
+  @override
+  num getValue(C category) => _data[category];
+
+  /**
+   * add or update the value of a specified category.
+   * [category] the pie category.
+   * [value] the pie value.
+   */
+  @override
+  void setValue(C category, num value) {
+    _data[category] = value;
+  }
+
+  // Sets color
+  @override
+  String getValueColor(C category) => _color[category];
 
   // Sets color of provided [series]
-  void setValueColor(String category, String color);
+  @override
+  void setValueColor(C category, String color) {
+    _color[category] = color;
+  }
 
+  /**
+   * Clear all data
+   */
+  @override
+  void clear() {
+    super.clear();
+
+    _data.clear();
+    _color.clear();
+  }
 }
 
 abstract class PieModel extends SingleValueCategoryModel {
 
 }
 
-abstract class DonutModel extends ChartModel {
+abstract class DonutModel<S extends Comparable, C extends Comparable> extends AbstractChartModel<S, C> {
 
-  void addSerie(String series, SingleValueCategoryModel model, {
+
+  /**
+   * Get (series) pairs of this chart data model. The returned
+   */
+  List<S> get keys;
+
+  /**
+   * Get value of the specified categoryModel
+   */
+  SingleValueCategoryModel<C> getValue(S series);
+
+  void addSeries(S series, SingleValueCategoryModel<C> model, {
     String size, String innerSize});
+}
 
+class DefaultDonutModel<S extends Comparable, C extends Comparable> extends DonutModel<S, C> {
+  final _data = <S, SingleValueCategoryModel<C>>{};
+
+  /**
+   * Get (series) pairs of this chart data model. The returned
+   */
+  List<S> get keys => _data.keys.toList();
+
+  /**
+   * Get value of the specified categoryModel
+   */
+  SingleValueCategoryModel<C> getValue(S series) => _data[series];
+
+  @override
+  void addSeries(S series, SingleValueCategoryModel<C> model, {
+    String size, String innerSize}) {
+    _data[series] = model;
+
+    setSeriesStyle(series, SeriesStyle.size, size);
+    setSeriesStyle(series, SeriesStyle.innerSize, innerSize);
+  }
+
+  /**
+   * Get all series as a collection.
+   */
+  List<S> get series {
+    return _data.keys;
+  }
+
+  /**
+   * Get categories as a collection.
+   */
+  List<C> get categories => <C>[];
+
+  /**
+   * Clear all data
+   */
+  void clear() {
+    super.clear();
+
+    _data.clear();
+  }
 }
 
 abstract class CategoryModel<S extends Comparable, C extends Comparable> extends ChartModel<S, C> {
@@ -83,19 +233,11 @@ abstract class CategoryModel<S extends Comparable, C extends Comparable> extends
    * [value] - the value
    */
   void setValue(S series, C category, num value);
-
-  // Returns color of provided [series]
-  String getSeriesColor(S series);
-
-  // Sets color of provided [series]
-  void setSeriesColor(S series, String color);
-
 }
 
-class DefaultCategoryModel<S extends Comparable, C extends Comparable> implements CategoryModel<S, C> {
+class DefaultCategoryModel<S extends Comparable, C extends Comparable> extends AbstractChartModel<S, C> implements CategoryModel<S, C> {
 
-  var _data = new LinkedHashMap<S, Map<C, num>>();
-  var _color = <S, String>{};
+  var _data = <S, Map<C, num>>{};
 
   @override
   List<List> get keys {
@@ -149,7 +291,7 @@ class DefaultCategoryModel<S extends Comparable, C extends Comparable> implement
 
   @override
   void setValue(S series, C category, num value) {
-    Map<C, num> map = _data[series] ?? new LinkedHashMap<C, num>();
+    Map<C, num> map = _data[series] ?? <C, num>{};
     map[category] = value;
     _data[series] = map;
 
@@ -157,18 +299,11 @@ class DefaultCategoryModel<S extends Comparable, C extends Comparable> implement
   }
 
   @override
-  String getSeriesColor(S series)
-    => _color[series];
-
-  @override
-  void setSeriesColor(S series, String color) {
-    _color[series] = color;
-  }
-
-  @override
   void clear() {
-    _data.clear();
-    _color.clear();
+    super.clear();
+    
+    _seriesStyle.clear();
+
     _categories = null;//clear cache
   }
 }