+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+**0.1.1**
+
+* Add implementation of DonutChart, AreaChart
+* Add DefaultSingleValueCategoryModel, DonutModel
+
 **0.1.0**
 
 * Add ColumnChart, DonutChart, AreaChart